modifiche reschedule, cascade tolte

+++ meteoCalProj/src/main/java/EJB/EventManagerImpl.java
@@ -225,6 +225,10 @@ public class EventManagerImpl implements EventManager {
         //aggiorno inizio evento
         if (oldEvent.getStartDateTime() != event.getStartDateTime()) {
             oldEvent.setStartDateTime(event.getStartDateTime());
+
+            //aggiorno il tempo al nuovo giorno
+            weatherManager.updateWeather(oldEvent);
+
             changed = true;
         }
         //aggiorno fine evento
@@ -434,23 +438,29 @@ public class EventManagerImpl implements EventManager {
 
     @Override
     public boolean deleteEvent(Event event) {
-        try {
-            event = database.find(Event.class, event.getId());
-            if (event instanceof PublicEvent) {
-                PublicEvent publicEvent = (PublicEvent) event;
-                notificationManager.createNotifications(publicEvent.getGuests(),
-                        event, NotificationType.EVENT_CANCELLED, false);
+        //TODO è sbagliata cancella i calendari dell'utente
+        if (event != null) {
+            try {
+                event = database.find(Event.class, event.getId());
+                CalendarModel cal;
+                if (event instanceof PublicEvent) {
+                    PublicEvent publicEvent = (PublicEvent) event;
+                    notificationManager.createNotifications(
+                            publicEvent.getGuests(),
+                            event, NotificationType.EVENT_CANCELLED, true);
+                }
+                notificationManager.createNotifications(event.getInvitee(),
+                        event,
+                        NotificationType.EVENT_CANCELLED, true);
+                database.remove(event);
+                return true;
+            } catch (IllegalArgumentException e) {
+                logger.log(LoggerLevel.DEBUG, "Evento: {0} non trovato",
+                        event.getId());
+                return false;
             }
-            notificationManager.createNotifications(event.getInvitee(), event,
-                    NotificationType.EVENT_CANCELLED, false);
-            database.remove(event);
-            return true;
-        } catch (IllegalArgumentException e) {
-            logger.log(LoggerLevel.DEBUG, "Evento: {0} non trovato",
-                    event.getId());
-            return false;
         }
-
+        return false;
     }
 
     /**
